+++ elements.js
@@ -1,0 +1,4 @@
+export const formFight = document.querySelector('.control');
+export const arenas = document.querySelector('.arenas');
+export const fightButton = formFight.querySelector('.button');
+export const chat = document.querySelector('.chat');
