+++ players.js
@@ -1,0 +1,27 @@
+import { changeHP, elHP, renderHP, attack } from "./hp.js";
+
+export const player1 = {
+    player: 1,
+    name: 'Scorpion',
+    hp: 100,
+    img: 'http://reactmarathon-api.herokuapp.com/assets/scorpion.gif',
+    weapon: ['lasso', 'knife', 'lash'],
+    damage: 0,
+    attack,
+    changeHP,
+    elHP,
+    renderHP,
+};
+
+export const player2 = {
+    player: 2,
+    name: 'Subzero',
+    hp: 100,
+    img: 'http://reactmarathon-api.herokuapp.com/assets/subzero.gif',
+    weapon: ['gun', 'fork', 'ray'],
+    damage: 0,
+    attack,
+    changeHP,
+    elHP,
+    renderHP,
+};
