+++ utils.js
@@ -1,0 +1,1 @@
+export const getRandom = (num) => {Math.ceil(Math.random() * num)}
